+++ R/aes_contrast.R
@@ -1,5 +1,6 @@
 #' Get contrast
 #'
+#' @param ... Provided to force user argument naming etc.
 #' @param fill A fill aesthetic from which to determine the colour scale for contrast.
 #' @param dark A dark colour. Defaults to `"black"`.
 #' @param light A light colour. Defaults to `"white"`.
@@ -20,9 +21,9 @@ get_contrast <- function(fill, dark = "black", light = "white") {
 #'
 #' @description A colour aesthetic for annotation that automatically contrasts with fill. Can be spliced into [ggplot2::aes] with [rlang::!!!].
 #'
-#' @param mode_family The mode family to optimise light and dark colours for. Options are "light", "grey" or "dark".
-#' @param dark A dark colour. If NULL, uses `mode_family` optimised colour.
-#' @param light A light colour. If NULL, uses `mode_family` optimised colour.
+#' @param ... Provided to force user argument naming etc.
+#' @param dark A dark colour.
+#' @param light A light colour.
 #'
 #' @return An aesthetic
 #' @export
@@ -64,25 +65,12 @@ get_contrast <- function(fill, dark = "black", light = "white") {
 #'     mode = dark_mode_r(),
 #'   ) +
 #'   geom_text(
-#'     mapping = aes(label = n, !!!aes_contrast("dark")),
+#'     mapping = aes(label = n, !!!aes_contrast(dark = darkness[3], light = darkness[1])),
 #'     position = position_dodge2(width = 0.75, preserve = "single"),
 #'     vjust = 1.33,
 #'     show.legend = FALSE,
 #'   )
-aes_contrast <- function(mode_family = "light", dark = NULL, light = NULL) {
-
-  if (mode_family == "light") {
-    if (rlang::is_null(dark)) dark <- "#121b24ff"
-    if (rlang::is_null(light)) light <- "#ffffffff"
-  }
-  else if (mode_family == "grey") {
-    if (rlang::is_null(dark)) dark <- "#121b24ff"
-    if (rlang::is_null(light)) light <- "#f6f8faff"
-  }
-  else if (mode_family == "dark") {
-    if (rlang::is_null(dark)) dark <- "#050d1bff"
-    if (rlang::is_null(light)) light <- "#c8d7dfff"
-  }
+aes_contrast <- function(..., dark = "#121B24FF", light = "#FFFFFFFF") {
 
   ggplot2::aes(
     colour = ggplot2::after_scale(