+++ R/grey_mode.R
@@ -7,6 +7,7 @@
 #' * `grey_mode_b()` with legend on bottom
 #' * `grey_mode_n()` with no legend
 #'
+#' @param ... Provided to force user argument naming etc.
 #' @param base_size The base size of the text theme element. Defaults to 11.
 #' @param base_family The base family of the text theme element. Defaults to "".
 #' @param base_colour The base colour of the text theme element.
@@ -28,7 +29,6 @@
 #' @param legend_ticks_linewidth The linewidth of the legend.ticks theme element.
 #' @param legend_ticks_length The legend.ticks.length theme element.
 #' @param orientation The orientation of the plot. Either "x" or "y". Defaults to NULL. Not intended for use with the mode argument of gg_* functions.
-#' @param ... Provided to support trailing commas only.
 #'
 #' @return A ggplot theme.
 #' @export
@@ -72,19 +72,20 @@
 #'   )
 #'
 grey_mode_r <- function (
+    ...,
     base_size = 11,
     base_family = "",
-    base_colour = "#121b24ff",
-    axis_line_colour = "#121b24ff",
+    base_colour = "#121B24FF",
+    axis_line_colour = "#121B24FF",
     axis_line_linewidth = 0.33,
     axis_ticks_colour = axis_line_colour,
     axis_ticks_linewidth = axis_line_linewidth,
     axis_ticks_length_x = grid::unit(base_size / 3, "pt"),
     axis_ticks_length_y = grid::unit(base_size / 4, "pt"),
-    panel_grid_colour = "#f6f8faff",
+    panel_grid_colour = "#F6F8FAFF",
     panel_grid_linewidth = 1.33,
-    panel_background_fill ="#fcfdfeff",
-    plot_background_fill = "#f6f8faff",
+    panel_background_fill ="#FCFDFEFF",
+    plot_background_fill = "#F6F8FAFF",
     legend_axis_line_colour = plot_background_fill,
     legend_axis_line_linewidth = 0.33,
     legend_background_fill = plot_background_fill,
@@ -92,15 +93,12 @@ grey_mode_r <- function (
     legend_ticks_colour = legend_axis_line_colour,
     legend_ticks_linewidth = legend_axis_line_linewidth,
     legend_ticks_length = ggplot2::rel(c(0.175, 0)),
-    orientation = NULL,
-    ...
-) {
+    orientation = NULL) {
 
   flex_mode_r(
     base_size = base_size,
     base_family = base_family,
     base_colour = base_colour,
-
     base_face = "plain",
     plot_title_size = ggplot2::rel(1.1),
     plot_title_family = base_family,
@@ -133,27 +131,26 @@ grey_mode_r <- function (
     legend_ticks_colour = legend_ticks_colour,
     legend_ticks_linewidth = legend_ticks_linewidth,
     legend_ticks_length = legend_ticks_length,
-    orientation = orientation,
-    ...
-  )
+    orientation = orientation)
 }
 
 #' @rdname grey_mode_r
 #' @export
 grey_mode_t <- function (
+    ...,
     base_size = 11,
     base_family = "",
-    base_colour = "#121b24ff",
-    axis_line_colour = "#121b24ff",
+    base_colour = "#121B24FF",
+    axis_line_colour = "#121B24FF",
     axis_line_linewidth = 0.33,
     axis_ticks_colour = axis_line_colour,
     axis_ticks_linewidth = axis_line_linewidth,
     axis_ticks_length_x = grid::unit(base_size / 3, "pt"),
     axis_ticks_length_y = grid::unit(base_size / 4, "pt"),
-    panel_grid_colour = "#f6f8faff",
+    panel_grid_colour = "#F6F8FAFF",
     panel_grid_linewidth = 1.33,
-    panel_background_fill ="#fcfdfeff",
-    plot_background_fill = "#f6f8faff",
+    panel_background_fill ="#FCFDFEFF",
+    plot_background_fill = "#F6F8FAFF",
     legend_axis_line_colour = plot_background_fill,
     legend_axis_line_linewidth = 0.33,
     legend_background_fill = plot_background_fill,
@@ -161,15 +158,12 @@ grey_mode_t <- function (
     legend_ticks_colour = legend_axis_line_colour,
     legend_ticks_linewidth = legend_axis_line_linewidth,
     legend_ticks_length = ggplot2::rel(c(0.175, 0)),
-    orientation = NULL,
-    ...
-) {
+    orientation = NULL) {
 
   flex_mode_t(
     base_size = base_size,
     base_family = base_family,
     base_colour = base_colour,
-
     base_face = "plain",
     plot_title_size = ggplot2::rel(1.1),
     plot_title_family = base_family,
@@ -202,27 +196,26 @@ grey_mode_t <- function (
     legend_ticks_colour = legend_ticks_colour,
     legend_ticks_linewidth = legend_ticks_linewidth,
     legend_ticks_length = legend_ticks_length,
-    orientation = orientation,
-    ...
-  )
+    orientation = orientation)
 }
 
 #' @rdname grey_mode_r
 #' @export
 grey_mode_b <- function (
+    ...,
     base_size = 11,
     base_family = "",
-    base_colour = "#121b24ff",
-    axis_line_colour = "#121b24ff",
+    base_colour = "#121B24FF",
+    axis_line_colour = "#121B24FF",
     axis_line_linewidth = 0.33,
     axis_ticks_colour = axis_line_colour,
     axis_ticks_linewidth = axis_line_linewidth,
     axis_ticks_length_x = grid::unit(base_size / 3, "pt"),
     axis_ticks_length_y = grid::unit(base_size / 4, "pt"),
-    panel_grid_colour = "#f6f8faff",
+    panel_grid_colour = "#F6F8FAFF",
     panel_grid_linewidth = 1.33,
-    panel_background_fill ="#fcfdfeff",
-    plot_background_fill = "#f6f8faff",
+    panel_background_fill ="#FCFDFEFF",
+    plot_background_fill = "#F6F8FAFF",
     legend_axis_line_colour = plot_background_fill,
     legend_axis_line_linewidth = 0.33,
     legend_background_fill = plot_background_fill,
@@ -230,15 +223,12 @@ grey_mode_b <- function (
     legend_ticks_colour = legend_axis_line_colour,
     legend_ticks_linewidth = legend_axis_line_linewidth,
     legend_ticks_length = ggplot2::rel(c(0.175, 0)),
-    orientation = NULL,
-    ...
-) {
+    orientation = NULL) {
 
   flex_mode_b(
     base_size = base_size,
     base_family = base_family,
     base_colour = base_colour,
-
     base_face = "plain",
     plot_title_size = ggplot2::rel(1.1),
     plot_title_family = base_family,
@@ -271,27 +261,26 @@ grey_mode_b <- function (
     legend_ticks_colour = legend_ticks_colour,
     legend_ticks_linewidth = legend_ticks_linewidth,
     legend_ticks_length = legend_ticks_length,
-    orientation = orientation,
-    ...
-  )
+    orientation = orientation)
 }
 
 #' @rdname grey_mode_r
 #' @export
 grey_mode_n <- function (
+    ...,
     base_size = 11,
     base_family = "",
-    base_colour = "#121b24ff",
-    axis_line_colour = "#121b24ff",
+    base_colour = "#121B24FF",
+    axis_line_colour = "#121B24FF",
     axis_line_linewidth = 0.33,
     axis_ticks_colour = axis_line_colour,
     axis_ticks_linewidth = axis_line_linewidth,
     axis_ticks_length_x = grid::unit(base_size / 3, "pt"),
     axis_ticks_length_y = grid::unit(base_size / 4, "pt"),
-    panel_grid_colour = "#f6f8faff",
+    panel_grid_colour = "#F6F8FAFF",
     panel_grid_linewidth = 1.33,
-    panel_background_fill ="#fcfdfeff",
-    plot_background_fill = "#f6f8faff",
+    panel_background_fill ="#FCFDFEFF",
+    plot_background_fill = "#F6F8FAFF",
     legend_axis_line_colour = plot_background_fill,
     legend_axis_line_linewidth = 0.33,
     legend_background_fill = plot_background_fill,
@@ -299,15 +288,12 @@ grey_mode_n <- function (
     legend_ticks_colour = legend_axis_line_colour,
     legend_ticks_linewidth = legend_axis_line_linewidth,
     legend_ticks_length = ggplot2::rel(c(0.175, 0)),
-    orientation = NULL,
-    ...
-) {
+    orientation = NULL) {
 
   flex_mode_n(
     base_size = base_size,
     base_family = base_family,
     base_colour = base_colour,
-
     base_face = "plain",
     plot_title_size = ggplot2::rel(1.1),
     plot_title_family = base_family,
@@ -340,7 +326,6 @@ grey_mode_n <- function (
     legend_ticks_colour = legend_ticks_colour,
     legend_ticks_linewidth = legend_ticks_linewidth,
     legend_ticks_length = legend_ticks_length,
-    orientation = orientation,
-    ...
+    orientation = orientation
   )
 }