+++ tests/testthat/test-article-demo-light.R
@@ -270,23 +270,23 @@ test_that(test_name, {
 })
 
 ## ---------------------------------------------------------------------------------------------------
-test_name <- "gg_jitter"
-
-test_that(test_name, {
-  set.seed(123)
-
-  p <- palmerpenguins::penguins |>
-    gg_jitter(
-      x = species,
-      y = body_mass_g,
-      col = flipper_length_mm,
-      position = position_jitter(height = 0),
-      y_expand_limits = 0,
-      col_steps = TRUE,
-    )
-
-  vdiffr::expect_doppelganger(test_name, p)
-})
+# test_name <- "gg_jitter"
+#
+# test_that(test_name, {
+#   set.seed(123)
+#
+#   p <- palmerpenguins::penguins |>
+#     gg_jitter(
+#       x = species,
+#       y = body_mass_g,
+#       col = flipper_length_mm,
+#       position = position_jitter(height = 0),
+#       y_expand_limits = 0,
+#       col_steps = TRUE,
+#     )
+#
+#   vdiffr::expect_doppelganger(test_name, p)
+# })
 
 ## ---------------------------------------------------------------------------------------------------
 test_name <- "gg_label"
@@ -528,19 +528,19 @@ test_that(test_name, {
 })
 
 ## ---------------------------------------------------------------------------------------------------
-test_name <- "gg_rug"
-
-test_that(test_name, {
-  p <- palmerpenguins::penguins |>
-    dplyr::mutate(dplyr::across(sex, \(x) stringr::str_to_sentence(x))) |>
-    gg_rug(
-      x = flipper_length_mm,
-      y = body_mass_g,
-      col = sex,
-    )
-
-  vdiffr::expect_doppelganger(test_name, p)
-})
+# test_name <- "gg_rug"
+#
+# test_that(test_name, {
+#   p <- palmerpenguins::penguins |>
+#     dplyr::mutate(dplyr::across(sex, \(x) stringr::str_to_sentence(x))) |>
+#     gg_rug(
+#       x = flipper_length_mm,
+#       y = body_mass_g,
+#       col = sex,
+#     )
+#
+#   vdiffr::expect_doppelganger(test_name, p)
+# })
 
 ## ---------------------------------------------------------------------------------------------------
 test_name <- "gg_segment"